replacing CompetitionBlock with CompetitionBlockAltB and renaming accordingly
