+++ resources/assets/components/CompetitionBlock/CompetitionBlock.js
@@ -3,7 +3,7 @@ import React from 'react';
 import classnames from 'classnames';
 
 import Markdown from '../Markdown';
-import { BlockWrapper } from '../Block';
+import Block, { BlockTitle } from '../Block';
 import Byline from '../Byline';
 import LazyImage from '../LazyImage';
 import './competitionBlock.scss';
@@ -11,17 +11,13 @@ import './competitionBlock.scss';
 const DEFAULT_CONFIRMATION = `
 # You're signed up!
 
-You should receive an email shortly with more instructions.
+You should recieve an email shortly with more instructions.
 You can keep working on the campaign for now. I'm so excited to have you onboard!
 `;
 
 const CompetitionBlock = (props) => {
   const { content, photo, byline, joinCompetition, hasJoinedCompetition, hasPendingJoin,
     showConfirmation, campaignId, campaignRunId, checkForCompetition, experiment, convert } = props;
-
-  if (! content) {
-    return null;
-  }
 
   // If we already joined the competition & saw the confirmation message,
   // display nothing.
@@ -40,10 +36,7 @@ const CompetitionBlock = (props) => {
       disabled={hasPendingJoin}
       className={classnames('button', { 'is-loading': hasPendingJoin })}
       onClick={() => {
-        if (experiment) {
-          convert(experiment);
-        }
-
+        convert(experiment);
         joinCompetition(campaignId, campaignRunId);
       }}
     >join competition</button>
@@ -56,29 +49,30 @@ const CompetitionBlock = (props) => {
   ) : null;
 
   return (
-    <BlockWrapper title="Go above and beyond!" className="-default">
+    <Block className={classnames('-colorful')}>
+      <BlockTitle>Go above and beyond!</BlockTitle>
       <div className={classnames('competition-block', { 'is-confirmation': showConfirmation })}>
         <div className="clearfix">
-          <Markdown className={classnames({ 'is-success': showConfirmation })}>{ showConfirmation ? DEFAULT_CONFIRMATION : content }</Markdown>
+          <Markdown className={classnames('', { 'is-success': showConfirmation })}>{ showConfirmation ? DEFAULT_CONFIRMATION : content }</Markdown>
           { competitionPhoto }
         </div>
         { button }
         <Byline {...byline} />
       </div>
-    </BlockWrapper>
+    </Block>
   );
 };
 
 CompetitionBlock.propTypes = {
   content: PropTypes.string.isRequired,
-  convert: PropTypes.func,
+  convert: PropTypes.func.isRequired,
   photo: PropTypes.string,
   byline: PropTypes.shape({
     author: PropTypes.string.isRequired,
     jobTitle: PropTypes.string.isRequired,
     avatar: PropTypes.string.isRequired,
   }).isRequired,
-  experiment: PropTypes.string,
+  experiment: PropTypes.string.isRequired,
   joinCompetition: PropTypes.func.isRequired,
   checkForCompetition: PropTypes.func.isRequired,
   hasJoinedCompetition: PropTypes.bool.isRequired,
@@ -90,8 +84,6 @@ CompetitionBlock.propTypes = {
 
 CompetitionBlock.defaultProps = {
   photo: null,
-  experiment: null,
-  convert: () => {},
 };
 
 export default CompetitionBlock;
