+++ resources/assets/components/CompetitionBlock/CompetitionBlockContainer.js
