+++ resources/assets/components/CompetitionBlock/index.js
@@ -1,7 +1,3 @@
 export CompetitionBlock from './CompetitionBlock';
 
-export CompetitionContainer from './CompetitionContainer';
-
-export CompetitionBlockAltB from './CompetitionBlockAltB';
-
-export CompetitionContainerAltB from './CompetitionContainerAltB';
+export default from './CompetitionBlockContainer';
